refactor: move shared components to shared
This allows easier distinction between "core components" and feature
components. Also fixes all paths.
