--- src/components/button/wrapper/buttonWrapper.css.js
@@ -1,9 +1,0 @@
-import styled from 'styled-components';
-
-export const Container = styled.div`
-  display: flex;
-  flex-flow: row wrap;
-  button + button {
-    margin-left: 1rem;
-  }
-`;
