--- src/components/io-example/io-example.css.js
@@ -1,13 +1,0 @@
-import styled from 'styled-components';
-
-export const Container = styled.div`
-  padding: 4rem;
-  margin: 4rem;
-  border-radius: 10px;
-  position: relative;
-  text-align: center;
-  color: #fff;
-  transition: background-color 0.3s ease;
-
-  background-color: ${({ isVisible }) => (isVisible ? ' #f00' : ' #333')};
-`;
