+++ src/components/shared/text/text.js
