+++ src/pages/404.js
@@ -1,5 +1,5 @@
 import React from 'react';
-import Box from 'components/box';
+import Box from 'components/shared/box';
 import Layout from 'components/layout';
 
 const NotFound = () => (
